feat(ko): Implement file watch dependencies
Add support for the `dependencies` section in the ko builder config.

Use `**/*.go` instead of `.` as the default file watch pattern for the
ko builder. As @tejal29
[points out](https://github.com/GoogleContainerTools/skaffold/pull/6617#discussion_r719804744),
this prevents unintended image rebuilds based on changes to files in
`.git/`, `.idea/`, and `*.swap` files and similar.

For this pattern to be effective, the changes in #6605 are required.

**Tracking:** #6041
**Related:** #6605