feat(build/docker): support env as secret source
Used in buildkit mode, allow users to specify secrets provided via env vars to
Skaffold (and hence docker/buildkit) in addition to file sources ("src").

I am setting "src" and "env" as mutually exclusive here in skaffold, however
docker currently does not enforce this (but "env" takes precedence regardless
of the order if both specified).

If we receive issues about this, we can relax
it in a backwards-compatible way.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>