fix: Wait for context cancel in k8s pod watcher (#6643)
Make the Kubernetes pod watcher context aware, and select on
`context.Done()`. This means we can stop waiting, and acting on, pod
events when the context has been cancelled.

Remove waiting on `context.Done()` in the Kubernetes log aggregator,
container manager, and pod port forwarder. This is to eliminate the
chance that the pod watcher sends a `PodEvent` on a channel without a
waiting receiver.

This should help avoid deadlocks that can occur when pod watcher event
receivers stop reading from the channel that they've registered with the
pod watcher.

We still close the channels on the receiver side, which could increase
the chances of regression and re-occurrence of this issue.

Also uses a RWMutex in the pod watcher, though we could move this change
to a separate commit.

Fixes: #6424