Update _index.md
Adding the solution to errors couldn't find manifest for 'skaffold' when installing with scoop. This error is due to lack of extras bucket when scoop is being installed. Default bucket that comes with scoop is main. In the absence of the extras bucket, users won't be able to install skaffold and couldn't find manifest for 'skaffold' error will be encountered