docs(ko): Templating in `flags` and `ldflags`
Document how environment variable templating of `flags` and `ldflags` is
handled by `ko` (rather than by Skaffold).

Keeping this consistent with `ko` simplifies the transition to Skaffold
for existing `ko` users. The implementation in `ko` also matches that of
GoReleaser.

Note that this is inconsistent with other templating support in Skaffold.

Also clarify the ko builder support for the `.ko.yaml` config file.

Tracking: #6041