feat: ignore vim *.swp files in hot-reload example
It is not obvious that `skaffold dev` will rebuild the image
even when a file that is not mapped in `skaffold.yaml`'s
manual sync config is added or changed. `skaffold dev` instead
respects `.dockerignore` as I discovered
[in this issue](https://github.com/GoogleContainerTools/skaffold/issues/3124).

Add `**/*.swp` to both node and python directories in the hot-reload
example so that this example is more likely to work out-of-the-box
for new users.

Signed-off-by: David Xia <dxia@spotify.com>