fix(ko): Debug port forwarding for `ko://` images
This change ensures that images with the `ko://` prefix have the `dlv`
debug port forwarded by default when running `skaffold debug`.

The image build process sanitizes the image name by removing the `ko://`
prefix and lowercasing the Go import path. This means that the built
image name doesn't match the placeholder values in the Kubernetes
manifest files.

This change sanitizes the image placeholder values from the Kubernetes
resource files when matching with the built image names, which in turn
means that Skaffold sets up port forwarding.

Fixes: #6995