fix(ko): Default repo for pushing `ko://` images
When using the `ko://` scheme prefix for the Skaffold image name, the
rest of the image name, following the prefix, is a Go import path.

This import path maps to the Git repository (typically), and won't match
an image registry name (e.g., `ko://github.com/org/repo` vs
`gcr.io/project_id`).

With this change, if _all_ of the following are true, the build results
in an error that instructs the user to set the default repo:

- the image name in `skaffold.yaml` uses the `ko://` prefix; and
- the Skaffold default repo is _not_ set; and
- the image is being pushed to a registry (rather than sideloaded to a
  Docker daemon).

Fixes: #6933
Tracking: #6041