feat(ko): GCB remote builds for ko builder
Enables remote builds on Google Cloud Build when using the Skaffold `ko`
builder.

Since Skaffold embeds ko as a module, the Cloud Build step uses the
`skaffold` binary to build and push the image. This means that we don't
need to create and maintain a separate `ko` builder image.

Tracking: #7131
Related: #7258