refactor: use exclude directives instead of replace directives for pinning
The 'replace' directive is very heavy-handed: it replaces nodes in the
dependency graph with different source code. That distorts their
apparent transitive dependencies, and makes a module that uses
'exclude' directives much more difficult for downstream consumers to
use even if they aren't using the parts of the module that depend on
the replaced versions.

In contrast, the 'exclude' directive knocks out specific version
requirements from the module graph without otherwise affecting the
dependencies of any remaining module in the graph.

'exclude' directives also make it clearer why particular pins are
necessary: normally only the 'require' directive is needed to select
desired versions (since they are not upgraded implicitly), so having
'exclude' blocks for specific conflicting requirements more clearly
shows which of the transitive dependencies are believed to be
spurious.