docs: Tutorial: Go coverage profiles for e2e tests
New tutorial that shows how to use Skaffold post-deploy hooks to collect
coverage profiles from integration tests, often called end-to-end tests,
of Go applications.

This takes advantage of new functionality in Go 1.20:
https://go.dev/testing/coverage/

I couldn't find a way to include files from the `examples` directory in
the tutorial Markdown, so instead of duplicating files from the example
in the `docs-v2` directory, the tutorial instead points readers to the
files in the `examples` directory.