fix: Replace Kustomize field `patches` in examples
This change replaces the obsolete use of the Kustomize `patches` field
in the Skaffold examples with `patchesStrategicMerge`.

Context:
Kustomize v5 removed support for `patches` being an array of strings.

The `patchesStrategicMerge` field introduced in Kustomize v3 retains the
previous behavior of the `patches`.

The `patches` field can still be used in v5 with an array of `patch`
objects.

Additional information:
- https://github.com/kubernetes-sigs/kustomize/pull/4911
- https://github.com/kubernetes-sigs/kustomize/releases/tag/kustomize%2Fv5.0.0