fix(ko): Ko builder push vs load behavior
Ko v0.13.0 introduced a change to the handling of image names when
loading images into a Docker daemon (instead of pushing to a registry).

Specifically, ko v0.13.0 assumes that if the image repo name matches
the
[`LocalDomain` field](https://github.com/ko-build/ko/blob/v0.13.0/pkg/commands/options/publish.go#L37),
the build image should be loaded into the Docker daemon, regardless of
the values of the
[`Push` and `Local` fields](https://github.com/ko-build/ko/blob/v0.13.0/pkg/commands/options/publish.go#L52-L56).
The implementation of this behavior is in
[ko's `makePublisher() function](https://github.com/ko-build/ko/blob/v0.13.0/pkg/commands/resolver.go#L189-L190).

This change ensures that Skaffold only sets the `LocalDomain` field for
images that should be loaded.

Fixes: #8760