feat: allow specifying a remote config dependency from Google Cloud Storage (#9057)
* feat: Introduce remote config dependency from Google Cloud Storage.

* feat: Adjust remote config dependency for Google Cloud Storage to be more inline with git.

Instead of copying the skaffold config from GCS the user specifies a source that signifies the bucket and directory structure to copy recursively. For example, specifying the source "gs://my-bucket/dir1/*" will copy all the files and subdirectories stored in "gs://my-bucket/dir1/". This requires the user to also specify the path to the skaffold config relative to the provided source. In the example mentioned, if the skaffold config was stored in "gs://my-bucket/dir1/configs/skaffold.yaml" then the path required would be "configs/skaffold.yaml".

* docs: Generate doc changes for GCS remote config dependency

* fix: unnecessary proto conversion fix

* fix: import order in couple files causing lint issues

* fix: only perform checks that disable gcs config syncs if the cache exists

* fix: lint issue related to generated proto

* undo generated proto changes since it broke all checks

* Remove redundant is remote dependency check