fix: Helm deployer was not working with variable templatingin chartPath
I believe the chartPath property is supposed to be a templatable field.

In my use I have helm charts in a separate git repo from the application where the skaffold.yaml is.
We defined a skaffold.env.template file, that we add to git and then on checkout we copy this to skaffold.env and modify as needed,
so the path to the location to the other repo with a helm chart is a variable. and this skaffold.env file is added to the .gitignore

We noticed that we were getting
listing files: issue walking releases: lstat {{.HELM_REPO}}/helm/app/chart/path: no such file or directory

This PR fixes this issue.