Bugfix: Only prune images if changed (#9787)
* fix: Only prune images if changed

I hope I understood this right...

*Description*
I observed that when using `skaffold dev`, when images were re-built because the dependencies defined in Skaffold changed (I use ko), but didn't result in a new image ID (because the contents hadn't changed - due to dependencies being too broad), then the previous image would be pruned. This means that images which aren't currently in use, like those used by k8s Jobs and CronJobs are removed, causing the deployment to fail.

*Solution*
Now we check whether the image ID of the existing artifact changed before removing it. Because if it's the same then we should keep it.

I tested this solution locally and seems to fix my problem.

* Add test for pruning behavior, including skipping if image ID is the same

* Fix linter and test issues

* Reorder imports