chore: fix the proto generation
It was using a very old version of Go which was, in turn, using a very
old version of Debian. That version of Debian hasn't been supported for
over a year, so trying to do an `apt-get upgrade` fails.

Upgrading go to fix that issue required upgrading a whole bunch of other
things. Exciting.